% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affy_2cells.R
\docType{data}
\name{affy_2cells}
\alias{affy_2cells}
\title{Gene expression data for samples with specific cell-type proportions}
\format{
A matrix
}
\source{
https://github.com/moyanre/smcgenedeconv
}
\usage{
affy_2cells
}
\description{
A matrix contaning GE data for samples with heart/brain cells
}
\keyword{datasets}
