% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_heart_spec_ge.R
\docType{data}
\name{brain_heart_spec_ge}
\alias{brain_heart_spec_ge}
\title{Cell-type specific GE data for brain/heart cells}
\format{
A matrix
}
\source{
https://github.com/moyanre/smcgenedeconv
}
\usage{
brain_heart_spec_ge
}
\description{
A matrix contaning average GE data for brain/heart cells
}
\keyword{datasets}
