% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_est_props.R
\name{combine_est_props}
\alias{combine_est_props}
\title{combine_est_props}
\usage{
combine_est_props(M_est, M)
}
\arguments{
\item{M_est}{a list containing matrices with deconvolution results}

\item{M}{a matrix of cell-type proportions}
}
\description{
combine_est_props
}
