% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mixture_samples.R
\name{create_mixture_samples}
\alias{create_mixture_samples}
\title{create_mixture_samples}
\usage{
create_mixture_samples(X, M)
}
\arguments{
\item{X}{cell-type specific GE data}

\item{M}{cell-type proportions of samples}
}
\value{
GE matrix of samples with with cell-type proportions given by M
}
\description{
create_mixture_samples
}
