% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{deconvolution}
\alias{deconvolution}
\title{deconvolution}
\usage{
deconvolution(X, Y, N = 40)
}
\arguments{
\item{X}{cell-type specific GE data}

\item{Y}{bulk GE data}

\item{N}{number of samples}
}
\value{

}
\description{
deconvolution
}
\examples{
deconvolution(X,Y,30)
}
