% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvg_selection.R
\name{hvg_selection}
\alias{hvg_selection}
\title{hvg_selection}
\usage{
hvg_selection(X, p = 0.05)
}
\arguments{
\item{X}{cell-type specific GE data}

\item{p}{include top (100*p)\% highly variable genes}
}
\value{
GE matrix with highly variable genes
}
\description{
hvg_selection
}
