% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsd_mse.R
\name{jsd_mse}
\alias{jsd_mse}
\title{jsd_mse}
\usage{
jsd_mse(M, M_est)
}
\arguments{
\item{M}{cell-type proportions}

\item{M_est}{deconvolution results}
}
\value{
matrix containing JSD between true and estimated proportions for each cell type.
}
\description{
jsd_mse
}
