% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mad.R
\name{mad}
\alias{mad}
\title{mad}
\usage{
mad(M, M_est)
}
\arguments{
\item{M}{cell-type proportions}

\item{M_est}{deconvolution results}
}
\value{
mean absolute difference (MAD) between true and estimated proportions
}
\description{
mad
}
