% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mad_plot.R
\name{mad_plot}
\alias{mad_plot}
\title{mad_plot}
\usage{
mad_plot(mad, size)
}
\arguments{
\item{mad}{vector of MAD values}

\item{size}{vector of sample size}
}
\value{
plot of MAD between true and estimated proportions for different sample sizes
}
\description{
mad_plot
}
