% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_rdirichlet.R
\name{prop_rdirichlet}
\alias{prop_rdirichlet}
\title{prop_rdirichlet}
\usage{
prop_rdirichlet(seed, n, k, types)
}
\arguments{
\item{seed}{seed set}

\item{n}{number of samples}

\item{k}{number of cell types}

\item{types}{cell types}
}
\description{
prop_rdirichlet
}
