% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\docType{data}
\name{proportions}
\alias{proportions}
\title{Cell-type proportions for samples in 'affy_2cells'}
\format{
A matrix
}
\source{
https://github.com/moyanre/smcgenedeconv
}
\usage{
proportions
}
\description{
A matrix containing cell-type proportions of brain/heart cells
}
\keyword{datasets}
