% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{scatter_plot}
\usage{
scatter_plot(M, M_est)
}
\arguments{
\item{M}{cell-type proportions}

\item{M_est}{deconvolution results}
}
\description{
scatter_plot
}
