% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_summary.R
\name{simulation_summary}
\alias{simulation_summary}
\title{simulation_summary}
\usage{
simulation_summary(t)
}
\arguments{
\item{t}{a vector of run times}
}
\value{
mean of t

standard deviation of t
}
\description{
simulation_summary
}
