% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_deconvolution.R
\name{smc_deconvolution}
\alias{smc_deconvolution}
\title{smc_deconvolution}
\usage{
smc_deconvolution(X, Y, M, N = 40, L = 1, num_cores = 1)
}
\arguments{
\item{X}{cell-type specific GE data}

\item{Y}{bulk GE data}

\item{N}{number of samples}

\item{L}{number of iterations}
}
\value{
a data frame with deconvolution results
}
\description{
smc_deconvolution
}
