% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{prop_plot}
\alias{prop_plot}
\title{prop_plot}
\usage{
prop_plot(X, Y, M, N = 40, L = 1, num_cores = 1)
}
\arguments{
\item{X}{cell-type specific GE data}

\item{Y}{bulk GE data}

\item{N}{number of samples}

\item{L}{number of iterations}
}
\description{
prop_plot
}
