% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\docType{data}
\name{proportions}
\alias{proportions}
\title{Cell-type proportions for samples in 'affy_2cells'}
\format{
A data frame
}
\source{
https://github.com/moyanre/smcgenedeconv
}
\usage{
proportions
}
\description{
A data frame containing cell-type proportions of heart/brain cells
}
\keyword{datasets}
