% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_deconvolution.R
\name{smc_deconvolution}
\alias{smc_deconvolution}
\title{smc_deconvolution}
\usage{
smc_deconvolution(Y, K, N = 40, L = 1, parallel = F, num_cores = 1)
}
\arguments{
\item{Y}{bulk GE data}

\item{K}{number of cell types}

\item{N}{number of samples}

\item{L}{number of iterations}

\item{parallel}{whether or not to use parallel computing}

\item{num_cores}{number of cores}
}
\value{
a data frame with deconvolution results

runtime of method
}
\description{
smc_deconvolution
}
